LayoutTests/fast/canvas/webgl/bug-32456.html

<html>
<head>
<link rel="stylesheet" href="../../js/resources/js-test-style.css"/>
<script src="../../js/resources/js-test-pre.js"></script>
<script src="resources/webgl-test.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>

<script>

description('Regression test for <a href="https://bugs.webkit.org/show_bug.cgi?id=32456">https://bugs.webkit.org/show_bug.cgi?id=32456</a> : <code>Errors in WebGL[T]Array set() and get() methods</code>');

debug("Testing WebGLFloatArray indexed getter / setter");
var array = new WebGLFloatArray([1.5]);
shouldBe("array[0]", "1.5");
shouldBe("array[0] = 2.5", "2.5");
shouldBe("array[0]", "2.5");

debug("Testing WebGLByteArray indexed getter / setter");
array = new WebGLByteArray(1);
shouldBe("array[0] = 127", "127");
shouldBe("array[0]", "127");
shouldBe("array[0] = -128", "-128");
shouldBe("array[0]", "-128");

debug("Testing WebGLUnsignedByteArray indexed getter / setter");
array = new WebGLUnsignedByteArray(1);
shouldBe("array[0] = 255", "255");
shouldBe("array[0]", "255");

debug("Testing WebGLShortArray indexed getter / setter");
array = new WebGLShortArray(1);
shouldBe("array[0] = 32767", "32767");
shouldBe("array[0]", "32767");
shouldBe("array[0] = -32768", "-32768");
shouldBe("array[0]", "-32768");

debug("Testing WebGLUnsignedShortArray indexed getter / setter");
array = new WebGLUnsignedShortArray(1);
shouldBe("array[0] = 65535", "65535");
shouldBe("array[0]", "65535");

debug("Testing WebGLIntArray indexed getter / setter");
array = new WebGLIntArray(1);
shouldBe("array[0] = -2147483648", "-2147483648");
shouldBe("array[0]", "-2147483648");
shouldBe("array[0] = 2147483647", "2147483647");
shouldBe("array[0]", "2147483647");

debug("Testing WebGLUnsignedIntArray indexed getter / setter");
array = new WebGLUnsignedIntArray(1);
shouldBe("array[0] = 4294967295", "4294967295");
shouldBe("array[0]", "4294967295");

successfullyParsed = true;

</script>
<script src="../../js/resources/js-test-post.js"></script>

</body>
</html>
